move rpc proxy into vnet api
