+++ deployables/app/app/routes.ts
@@ -162,10 +162,7 @@ export default [
     route('get-plan/:prefixedAddress', 'routes/system/get-plan.ts'),
   ]),
 
-  ...prefix('/vnet', [
-    index('routes/vnet/index.ts'),
-    route('rpc/:network/:slug', 'routes/vnet/rpc.$network.$slug.ts'),
-  ]),
+  ...prefix('/vnet', [index('routes/vnet.ts')]),
 
   ...prefix('/dev', [
     route('decode/:data', 'routes/dev/decode.tsx'),
