+++ deployables/app/app/routes/vnet.ts
@@ -1,5 +1,5 @@
 import type { ActionFunctionArgs } from 'react-router'
-import { createVnet } from '../../vnet/server/createVnet'
+import { createVnet } from '../vnet/server/createVnet'
 
 export async function action({ request }: ActionFunctionArgs) {
   return await createVnet(await request.json())
