+++ deployables/vnet-api/.env.sample
@@ -1,0 +1,3 @@
+  TENDERLY_ACCESS_KEY=
+  TENDERLY_USER=gnosisguild
+  TENDERLY_PROJECT=zodiac-pilot
