+++ deployables/vnet-api/eslint.config.mjs
@@ -1,0 +1,3 @@
+import zodiacEslintConfig from '@zodiac/eslint-config'
+
+export default zodiacEslintConfig
