+++ deployables/vnet-api/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@zodiac/vnet-api",
+  "version": "1.0.0",
+  "devDependencies": {
+    "@cloudflare/workers-types": "4.20250702.0",
+    "@zodiac/eslint-config": "workspace:*",
+    "@zodiac/typescript-config": "workspace:*",
+    "env-cmd": "10.1.0",
+    "eslint": "9.30.1",
+    "typescript": "5.8.3",
+    "wrangler": "4.22.0"
+  },
+  "private": true,
+  "scripts": {
+    "start": "env-cmd wrangler dev",
+    "deploy": "wrangler publish",
+    "check-types": "tsc --noEmit",
+    "lint": "eslint . --max-warnings=0"
+  },
+  "packageManager": "pnpm@10.12.4"
+}
