+++ deployables/vnet-api/src/types.ts
@@ -1,0 +1,5 @@
+export interface Env {
+  TENDERLY_ACCESS_KEY: string
+  TENDERLY_USER: string
+  TENDERLY_PROJECT: string
+}
